테스트 코드 작성 PR

+++ src/main/java/com/jay/shoppingmall/service/CartService.java
@@ -2,6 +2,7 @@ package com.jay.shoppingmall.service;
 
 import com.jay.shoppingmall.domain.cart.Cart;
 import com.jay.shoppingmall.domain.cart.CartRepository;
+import com.jay.shoppingmall.domain.image.Image;
 import com.jay.shoppingmall.domain.image.ImageRelation;
 import com.jay.shoppingmall.domain.image.ImageRepository;
 import com.jay.shoppingmall.domain.item.Item;
@@ -18,10 +19,7 @@ import com.jay.shoppingmall.dto.response.cart.CartPriceTotalResponse;
 import com.jay.shoppingmall.dto.response.item.ItemAndQuantityResponse;
 import com.jay.shoppingmall.dto.response.item.ItemOptionResponse;
 import com.jay.shoppingmall.dto.response.seller.SellerResponse;
-import com.jay.shoppingmall.exception.exceptions.AlreadyExistsException;
-import com.jay.shoppingmall.exception.exceptions.CartEmptyException;
-import com.jay.shoppingmall.exception.exceptions.ItemNotFoundException;
-import com.jay.shoppingmall.exception.exceptions.SellerNotFoundException;
+import com.jay.shoppingmall.exception.exceptions.*;
 import com.jay.shoppingmall.service.handler.FileHandler;
 import lombok.EqualsAndHashCode;
 import lombok.RequiredArgsConstructor;
@@ -47,19 +45,18 @@ public class CartService {
     private final ItemOptionRepository itemOptionRepository;
     private final SellerRepository sellerRepository;
 
-    //TODO 밑의 코드들로 리팩토링 하기
     public Map<SellerResponse, List<ItemAndQuantityResponse>> showCartItemsList(User user) {
+
         //유저를 기준으로 장바구니 물품 가져오기
         final List<Cart> carts = cartRepository.findByUser(user);
-//                .orElseThrow(() -> new UsernameNotFoundException("해당 유저가 없습니다"));
 
         Map<SellerResponse, List<ItemAndQuantityResponse>> sellerResponseListMap = new ConcurrentHashMap<>();
         final List<Seller> sellers = carts.stream().map(Cart::getItem).map(Item::getSeller).distinct().collect(Collectors.toList());
+
         //판매자를 기준으로 상품 정렬
         for (Seller seller : sellers) {
+            this.cartPricePerSeller(user, seller);
             final List<Cart> cartList = carts.stream().filter(cart -> cart.getItem().getSeller().equals(seller)).collect(Collectors.toList());
-            long itemTotalPricePerSeller = 0;
-            int itemTotalQuantityPerSeller = 0;
 
             //Key 판매자 생성
             final SellerResponse sellerResponse = SellerResponse.builder()
@@ -75,10 +72,13 @@ public class CartService {
                 itemAndQuantityResponses = new ArrayList<>();
             }
 
+            long itemTotalPricePerSeller = 0;
+            int itemTotalQuantityPerSeller = 0;
+
             //Value List<ItemAndQuantityResponse> 생성
             for (Cart cart : cartList) {
                 if (cart.getIsSelected()) {
-                    itemTotalPricePerSeller += cart.getItemOption().getItemPrice().getPriceNow() * cart.getQuantity();
+                    itemTotalPricePerSeller += cart.getPriceNow() * cart.getQuantity();
                     itemTotalQuantityPerSeller += cart.getQuantity();
                 }
                 itemAndQuantityResponses.add(ItemAndQuantityResponse.builder()
@@ -89,22 +89,35 @@ public class CartService {
                         .optionId(cart.getItemOption().getId())
                         .name(cart.getItem().getName())
                         .priceNow(cart.getItemOption().getItemPrice().getPriceNow())
-                        .image(fileHandler.getStringImage(imageRepository.findByImageRelationAndForeignId(ImageRelation.ITEM_MAIN, cart.getItem().getId())))
+//                        .image(fileHandler.getStringImage(imageRepository.findByImageRelationAndForeignId(ImageRelation.ITEM_MAIN, cart.getItem().getId())))
                         .zzim(cart.getItem().getZzim())
                         .quantity(cart.getQuantity())
                         .isSelected(cart.getIsSelected())
                         .build());
             }
+            setImage(cartList, itemAndQuantityResponses);
             //배송비 무료 여부 및 총 가격
-            sellerResponse.setShippingFee(shippingFeeCheck(seller.getId(), itemTotalPricePerSeller));
+            sellerResponse.setShippingFee(shippingFeeCheck(seller, itemTotalPricePerSeller));
             sellerResponse.updateTotal(itemTotalPricePerSeller, itemTotalQuantityPerSeller);
 
-//            if (itemTotalPricePerSeller == 0) {
-//                sellerResponse.setShippingFee(0);
-//            }
             sellerResponseListMap.put(sellerResponse, itemAndQuantityResponses);
         }
         return sellerResponseListMap;
+    }
+
+    private void setImage(final List<Cart> cartList, final List<ItemAndQuantityResponse> itemAndQuantityResponses) {
+        final List<Long> itemIds = cartList.stream().map(Cart::getItem).map(Item::getId).collect(Collectors.toList());
+        final List<Image> images = imageRepository.findByImageRelationAndForeignIdIn(ImageRelation.ITEM_MAIN, itemIds);
+
+        for (ItemAndQuantityResponse itemAndQuantityResponse : itemAndQuantityResponses) {
+            for (Image image : images) {
+                final String stringImage = fileHandler.getStringImage(image);
+                if (itemAndQuantityResponse.getItemId().equals(image.getForeignId())) {
+                    itemAndQuantityResponse.setImage(stringImage);
+                    break;
+                }
+            }
+        }
     }
 
     public void addOptionItemsToCart(final List<ItemOptionResponse> itemOptions, final User user) {
@@ -129,56 +142,26 @@ public class CartService {
         }
     }
 
-//    public Integer addItemToCart(CartRequest request, User user) {
-//        Item item = itemRepository.findById(request.getItemId())
-//                .orElseThrow(() -> new ItemNotFoundException("해당 상품이 존재하지않습니다"));
-//        ItemOption itemOption = itemOptionRepository.findByOption1AndOption2AndItemId(request.getOption1(), request.getOption2(), request.getItemId())
-//                .orElseThrow(() -> new ItemNotFoundException("해당 상품이 존재하지않습니다"));
-//
-//        Cart cart = cartRepository.findByUserAndItemAndItemOption(user, item, itemOption)
-//                .orElseThrow(() -> new AlreadyExistsException("해당 상품이 장바구니에 존재합니다"));
-//
-//        if (item.getStock() == 0) {
-//            throw new ItemNotFoundException("죄송합니다. 해당 상품은 품절되었습니다");
-//        }
-//        Integer addedQuantity = request.getQuantity();
-//        if (cart != null) {
-////            if (Objects.equals(cart.getItemOption().getId(), itemOption.getId())) {
-//            throw new AlreadyExistsException("이미 상품이 장바구니에 존재합니다");
-//            //            addedQuantity = quantity + cart.getQuantity();
-//            //            cart.manipulateQuantity(addedQuantity);
-//            //            cartRepository.save(cart);
-////            }
-//        } else {
-//            Cart newCart = Cart.builder()
-//                    .user(user)
-//                    .item(item)
-//                    .itemOption(itemOption)
-//                    .quantity(addedQuantity)
-//                    .build();
-//            cartRepository.save(newCart);
-//        }
-//        return addedQuantity;
-//    }
-
-    //장바구니 전체 가격, 총 배송비, 개수
     public CartPriceTotalResponse cartPriceTotal(User user) {
-        final List<Cart> carts = cartRepository.findByUserAndIsSelectedTrue(user);
-//                .orElseThrow(() -> new CartEmptyException("장바구니가 비어있습니다"));
 
-        long cartTotalPrice = 0;
-        int cartTotalQuantity = 0;
-        int cartTotalShippingFee = 0;
+        final List<CartPricePerSellerResponse> cartPricePerSellerResponses = cartPricePerSeller(user);
 
-        //장바구니 판매자 기준 분류
-        final List<Seller> sellers = carts.stream().map(Cart::getItem).map(Item::getSeller).distinct().collect(Collectors.toList());
-        for (Seller seller : sellers) {
-            final CartPricePerSellerResponse cartPricePerSellerResponse = this.cartPricePerSeller(user, seller);
+        long cartTotalPrice = cartPricePerSellerResponses.stream().mapToLong(CartPricePerSellerResponse::getItemTotalPricePerSeller).sum();
+        int cartTotalQuantity = cartPricePerSellerResponses.stream().mapToInt(CartPricePerSellerResponse::getItemTotalQuantityPerSeller).sum();
+        int cartTotalShippingFee = cartPricePerSellerResponses.stream().mapToInt(CartPricePerSellerResponse::getItemShippingFeePerSeller).sum();
 
-            cartTotalPrice += cartPricePerSellerResponse.getItemTotalPricePerSeller();
-            cartTotalQuantity += cartPricePerSellerResponse.getItemTotalQuantityPerSeller();
-            cartTotalShippingFee += cartPricePerSellerResponse.getItemShippingFeePerSeller();
-        }
+        return CartPriceTotalResponse.builder()
+                .cartTotalPrice(cartTotalPrice)
+                .cartTotalQuantity(cartTotalQuantity)
+                .cartTotalShippingFee(cartTotalShippingFee)
+                .build();
+    }
+    public CartPriceTotalResponse cartPriceTotal(List<CartPricePerSellerResponse> cartPricePerSellerResponses) {
+
+        long cartTotalPrice = cartPricePerSellerResponses.stream().mapToLong(CartPricePerSellerResponse::getItemTotalPricePerSeller).sum();
+        int cartTotalQuantity = cartPricePerSellerResponses.stream().mapToInt(CartPricePerSellerResponse::getItemTotalQuantityPerSeller).sum();
+        int cartTotalShippingFee = cartPricePerSellerResponses.stream().mapToInt(CartPricePerSellerResponse::getItemShippingFeePerSeller).sum();
+
         return CartPriceTotalResponse.builder()
                 .cartTotalPrice(cartTotalPrice)
                 .cartTotalQuantity(cartTotalQuantity)
@@ -189,32 +172,67 @@ public class CartService {
     //판매자별 가격, 개수, 배송비
     public CartPricePerSellerResponse cartPricePerSeller(User user, Seller seller) {
         final List<Cart> carts = cartRepository.findByUserAndIsSelectedTrue(user);
-//                .orElseThrow(() -> new CartEmptyException("장바구니가 비어있습니다"));
 
-        final List<Cart> cartList = carts.stream().filter(cart -> cart.getItem().getSeller().equals(seller)).collect(Collectors.toList());
+        final List<Cart> cartList = carts.stream().filter(cart -> cart.getItemSeller().equals(seller)).collect(Collectors.toList());
         long cartTotalPricePerSeller = 0;
         int cartTotalQuantityPerSeller = 0;
 
         for (Cart cart : cartList) {
-            cartTotalPricePerSeller += cart.getItemOption().getItemPrice().getPriceNow() * cart.getQuantity();
-            cartTotalQuantityPerSeller += cart.getQuantity();
+            if (cart.getIsSelected()) {
+                cartTotalPricePerSeller += cart.getPriceNow() * cart.getQuantity();
+                cartTotalQuantityPerSeller += cart.getQuantity();
+            }
         }
         return CartPricePerSellerResponse.builder()
                 .sellerId(seller.getId())
                 .itemTotalPricePerSeller(cartTotalPricePerSeller)
                 .itemTotalQuantityPerSeller(cartTotalQuantityPerSeller)
-                .itemShippingFeePerSeller(shippingFeeCheck(seller.getId(), cartTotalPricePerSeller))
+                .itemShippingFeePerSeller(shippingFeeCheck(seller, cartTotalPricePerSeller))
                 .build();
+    }
+
+    public List<CartPricePerSellerResponse> cartPricePerSeller(User user) {
+        List<CartPricePerSellerResponse> cartPricePerSellerResponses = new ArrayList<>();
+        final List<Cart> carts = cartRepository.findByUser(user);
+
+        final List<Seller> sellers = carts.stream().map(Cart::getItem).map(Item::getSeller).distinct().collect(Collectors.toList());
+        for (Seller seller : sellers) {
+            long cartTotalPricePerSeller = 0;
+            int cartTotalQuantityPerSeller = 0;
+
+            final List<Cart> cartsPerSeller = carts.stream().filter(cart -> cart.getItemSeller().equals(seller)).collect(Collectors.toList());
+
+            for (Cart cart : cartsPerSeller) {
+                if (!cart.getIsSelected())
+                    continue;
+
+                cartTotalPricePerSeller += cart.getPriceNow() * cart.getQuantity();
+                cartTotalQuantityPerSeller += cart.getQuantity();
+            }
+
+            final Integer shippingFee = shippingFeeCheck(seller, cartTotalPricePerSeller);
+
+            final CartPricePerSellerResponse cartPricePerSellerResponse = CartPricePerSellerResponse.builder()
+                    .sellerId(seller.getId())
+                    .itemTotalPricePerSeller(cartTotalPricePerSeller)
+                    .itemTotalQuantityPerSeller(cartTotalQuantityPerSeller)
+                    .itemShippingFeePerSeller(shippingFee)
+                    .build();
+
+            cartPricePerSellerResponses.add(cartPricePerSellerResponse);
+        }
+        return cartPricePerSellerResponses;
     }
 
     public CartPriceResponse changeQuantity(final CartManipulationRequest request, final User user) {
         final Cart cart = cartRepository.findById(request.getCartId())
                 .orElseThrow(() -> new CartEmptyException("장바구니가 비어있습니다"));
-
         cart.setIsSelected(request.getIsSelected());
+        cart.manipulateQuantity(request.getCartQuantity());
 
-        final long cartManipulatedPrice = cart.manipulateQuantity(request.getCartQuantity()) * cart.getItemOption().getItemPrice().getPriceNow();
-        final CartPricePerSellerResponse cartPricePerSellerResponse = cartPricePerSeller(user, cart.getItem().getSeller());
+        final long cartManipulatedPrice = cart.manipulateQuantity(request.getCartQuantity()) * cart.getPriceNow();
+
+        final CartPricePerSellerResponse cartPricePerSellerResponse = cartPricePerSeller(user, cart.getItemSeller());
         final CartPriceTotalResponse cartPriceTotalResponse = cartPriceTotal(user);
 
         return CartPriceResponse.builder()
@@ -223,16 +241,23 @@ public class CartService {
                 .cartPriceTotalResponse(cartPriceTotalResponse)
                 .build();
     }
-    public CartPriceResponse cartSelect(final String check, final User user) {
-        List<Cart> carts = cartRepository.findByUser(user);
-//                .orElseThrow(() -> new UsernameNotFoundException("해당 유저가 없습니다"));
-        List<CartPricePerSellerResponse> cartPricePerSellerResponses = new ArrayList<>();
-
-        for (Cart cart : carts) {
-            cart.setIsSelected(check.equals("true"));
-            cartPricePerSellerResponses.add(cartPricePerSeller(user, cart.getItem().getSeller()));
+    public CartPriceResponse cartSelectAll(final String check, final User user) {
+        if (!check.equals("false") && !check.equals("true")) {
+            throw new NotValidException("잘못된 요청입니다");
         }
-        final CartPriceTotalResponse cartPriceTotalResponse = cartPriceTotal(user);
+        List<Cart> carts = cartRepository.findByUser(user);
+
+        if (check.equals("false")) {
+            for (Cart cart : carts) {
+                cart.setIsSelected(false);
+            }
+        } else {
+            for (Cart cart : carts) {
+                cart.setIsSelected(true);
+            }
+        }
+        final List<CartPricePerSellerResponse> cartPricePerSellerResponses = cartPricePerSeller(user);
+        final CartPriceTotalResponse cartPriceTotalResponse = cartPriceTotal(cartPricePerSellerResponses);
 
         return CartPriceResponse.builder()
                 .cartPricePerSellerResponses(cartPricePerSellerResponses)
@@ -268,10 +293,13 @@ public class CartService {
     public CartPriceResponse deleteCart(final CartManipulationRequest request, final User user) {
         final Cart cart = cartRepository.findById(request.getCartId())
                 .orElseThrow(() -> new CartEmptyException("장바구니가 비어있습니다"));
+        if (!cart.getUser().getId().equals(user.getId())) {
+            throw new UserNotFoundException("잘못된 접근입니다");
+        }
         cartRepository.delete(cart);
 
-        final CartPricePerSellerResponse cartPricePerSellerResponse = cartPricePerSeller(user, cart.getItem().getSeller());
-        final CartPriceTotalResponse cartPriceTotalResponse = cartPriceTotal(user);
+        final CartPricePerSellerResponse cartPricePerSellerResponse = this.cartPricePerSeller(user, cart.getItem().getSeller());
+        final CartPriceTotalResponse cartPriceTotalResponse = this.cartPriceTotal(user);
 
         return CartPriceResponse.builder()
                 .cartPricePerSellerResponse(cartPricePerSellerResponse)
@@ -281,7 +309,7 @@ public class CartService {
 
     public Integer getTotalQuantity(final User user) {
         List<Cart> carts = cartRepository.findByUser(user);
-//                .orElseThrow(() -> new UsernameNotFoundException("해당 유저가 없습니다"));
+
         Integer totalQuantity = 0;
         for (Cart cart : carts) {
             totalQuantity += cart.getQuantity();
@@ -289,10 +317,7 @@ public class CartService {
         return totalQuantity;
     }
 
-    public Integer shippingFeeCheck(Long sellerId, Long itemTotalPricePerSeller) {
-        final Seller seller = sellerRepository.findById(sellerId)
-                .orElseThrow(() -> new SellerNotFoundException("판매자가 존재하지 않습니다"));
-
+    public Integer shippingFeeCheck(Seller seller, Long itemTotalPricePerSeller) {
         //무료 배송 정책이 있으며 구매 가격이 그 이상일 때 무료
         if (seller.getShippingFeeFreePolicy() != null && itemTotalPricePerSeller >= seller.getShippingFeeFreePolicy()) {
             return 0;
